+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## dev ##
+
+### enhancements
+  * Array-like attributes. (by [@lest](https://github.com/lest))
+
 ## 0.4.0 (September 6, 2012) ##
 
 Legacy support was dropped. The following versions are supported: